[Eir] Phase 1 - Prerequisites - updated to MTA:BLUE revision 6169
The new core of MTA:Eir has been greatly improved. It has performance improvements for low-end CPUs (fibered resource loading). The stability and quality has been revised. New features like dynamic lighting and render callback monitoring have been added. New scripting functions have been planned that can be looked up at...

http://wiki.mtasa.com/wiki/MTA:Eir/New_Scripting_Functions

I will talk about the features in every commit that they are included in. For now I have to do the groundwork. So, let's go!

BUILD STATUS: not-ready

+++ forks/eir/Shared/sdk/UTF8.h
@@ -212,7 +212,7 @@ utf8_mbstowcs (const std::string & str)
             s += un;
             sn += un;
         }
-        unsigned int usedsize = ptr - buffer;
+        size_t usedsize = ptr - buffer;
         dassert ( usedsize < cCharacters );
         return std::wstring( buffer, usedsize );
     }
@@ -251,7 +251,7 @@ utf8_wcstombs (const std::wstring & wstr)
         {
             ptr += utf8_wctomb ((unsigned char*)ptr, wstr [i], 6 );
         }
-        unsigned int usedsize = ptr - buffer;
+        size_t usedsize = ptr - buffer;
         dassert ( usedsize < cBytes );
         return std::string( buffer, usedsize );
     }
